finalChanges
finalChanges

+++ Nemo/Nemo/Manager/production_lines.xaml.cs
@@ -23,13 +23,13 @@ namespace Nemo.Manager
 
         Database.AppLayer applayer;
         addLinePage addLinePage;
-        ManagerOptoins parent;
+        ManagerOptoins parentInstance;
         DataTable data;
-        public production_lines(ManagerOptoins p)
+        public production_lines(ManagerOptoins parentInstance)
         {
             InitializeComponent();
             applayer = Database.AppLayer.GetInstance();
-            parent = p;
+            this.parentInstance = parentInstance;
             getAllLines();
             addLinePage = new addLinePage();
         }
@@ -38,16 +38,35 @@ namespace Nemo.Manager
             data = applayer.getAllLines();
             prodLinesGrid.ItemsSource = data.DefaultView;        }
 
-        private void Image_MouseDown(object sender, MouseButtonEventArgs e)
+        private void addLine(object sender, MouseButtonEventArgs e)
         {
-            parent.moreInfo.Content = addLinePage;
+            MakeSound.MakeClick();
+            new AddProductionLine(parentInstance).Show();
+            parentInstance.Hide();
 
         }
 
         private void itm_delete_Click(object sender, RoutedEventArgs e)
         {
-            if (applayer.deleteLine(((DataRowView)prodLinesGrid.SelectedItem)["ID"].ToString()) >0)
-                MessageBox.Show("Line deleted succesfully!");
+            if (MessageBox.Show("Are you sure you want to delete this this?", "Are you sure Nemo?", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            {
+                MakeSound.MakeSent();
+                if (applayer.deleteLine(((DataRowView)prodLinesGrid.SelectedItem)["ID"].ToString()) > 0)
+                    MessageBox.Show("Line deleted succesfully!");
+                getAllLines();
+            }
+        }
+        
+        private void itm_update_Click(object sender, RoutedEventArgs e)
+        {
+            MakeSound.MakeClick();
+            new UpdateProductionLine(((DataRowView)prodLinesGrid.SelectedItem)["ID"].ToString(),parentInstance).Show();
+            parentInstance.Hide();
+        }
+
+        public void Refresh()
+        {
+            prodLinesGrid.ItemsSource = null;
             getAllLines();
         }
     }
